requirements.txt
pyvips====2.1.14
pandas
beautifulsoup4
Pillow